fixed cmake and package.xml for libvectormap moved headers into include/libvectormap since otherwise this otherwise can conflict with other files elsewhere.
