+++ ros/src/computing/perception/detection/lib/libvectormap/package.xml
@@ -7,6 +7,8 @@
   <license>BSD</license>
   <buildtool_depend>catkin</buildtool_depend>
   <build_depend>vector_map</build_depend>
+  <build_depend>eigen</build_depend>
   <run_depend>vector_map</run_depend>
+  <run_depend>eigen</run_depend>
   <export></export>
 </package>
